+++ Lib/Registry/YarnNodePackageManager.cs
@@ -113,7 +113,7 @@ namespace Lib.Registry
         void RunYarnWithParam(IDirectoryCache projectDirectory, string param)
         {
             var fullPath = projectDirectory.FullPath;
-            var project = TSProject.Get(projectDirectory, _diskCache, _logger);
+            var project = TSProject.Get(projectDirectory, _diskCache, _logger, null);
             project.LoadProjectJson(true);
             if (project.ProjectOptions.NpmRegistry != null)
             {